requirements.txt
pyserial==3.4
crc8==0.1.0